Branched out dev fee messages to single script
