+++ secondLevel/utils/secondLevelCustMsg.py
@@ -30,34 +30,6 @@ async def account_layer_selection_message(destination, level: int):
     await destination.send(embed=wallet_selection_info)
 
 
-async def dev_fee_option_notification(destination):
-    """
-    Prompt user if he is willing to give dev fee
-    """
-    dev_fee_info = Embed(title=":robot: Optional Dev Fee for Crypto Link Team:robot: ",
-                         color=Colour.lighter_gray())
-    dev_fee_info.add_field(name=f':money_with_wings: Dev fee :money_with_wings: ',
-                           value=f'If you would like to support us in what we are doing please answer with'
-                                 f'***__Yes__*** and system will ask you for custom amount.'
-                                 f'If you choose **__No__** than dev fee will be skipped.',
-                           inline=False)
-    await destination.send(embed=dev_fee_info)
-
-
-async def ask_for_dev_fee_amount(destination):
-    """
-    Prompt to user to provide dev fee
-    """
-    dev_fee_info = Embed(title=":robot: Dev Fee Amount/Value :robot: ",
-                         color=Colour.lighter_gray())
-    dev_fee_info.add_field(name=f':money_with_wings: Value :money_with_wings: ',
-                           value=f'Please provide amount in format 0.0000000 you are willing to contribute to '
-                                 f'development. Crypto Link will than automatically append Payment operation to the '
-                                 f'current transactions.',
-                           inline=False)
-    await destination.send(embed=dev_fee_info)
-
-
 async def send_new_account_information(ctx, details: dict):
     """
     Information to user who has registered for second level wallet
